passim/requirements.txt
# Requirements for PASSIM
# Date: 28/aug/2023
#        2/okt/2023
#       24/jan/2024
# 
# Python version this works for: 3.10.12 / 3.8
#
# For a better overview of dependencies: install pipdeptree

django==4.1                 # OLD django==2.2.28
django-extensions==3.1.0
django-redis==5.2.0         # OLD: django-redis==4.11.0
django-select2==7.10.1      # OLD: django-select2==6.3.1
docopt==0.6.2
lxml==4.6.1
Markdown==3.3.4
openpyxl==2.6.2
PyGraph==0.2.1
xmltodict==0.12.0
# chardet - not needed
# demjson - overtaken by the built-in JSON (import json)
unidecode==1.3.7

# Note: the correct numpy version must be use for a couple of packages
numpy==1.22.1
  pandas==2.0.3
  scipy==1.9.3
  scikit-learn==1.3.0
# Note: 'import sklearn' is no longer possible as separate package
#       'from sklearn import XXX' works with just 'import scikit-learn'
  
# Note: pillow might install a version of reportlab too new  
pillow==10.2.0          # OLD: pillow==6.2.2
  reportlab==3.5.68

# Note: requests are needed anyway, so we pick the version
#       as well as the pyparsing an pyzotero going along with it
# Note: pyzotero will also install e.g. bibtexparser, pathlib and others
requests==2.26.0
  pyparsing==2.4.7
    pydotplus==2.0.2
    Pyzotero==1.5.10  

# Note: this is used by the documentation .rst files
#       (it does install a lot of packages)
sphinx==4.2.0

# Needed for the 'plugin' app:
django_plotly_dash==2.2.2
dash-bootstrap-templates==1.1.2
umap-learn==0.5.5